api_mongo/requirements.txt
Flask==1.1.1
Flask-PyMongo==2.3.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==1.0.0
click==7.1.1
itsdangerous==1.1.0
pymongo==3.10.1
python-dotenv==0.12.0
typing==3.7.4.1
